+++ src/main/java/dev/dahye/lotto/util/DoubleUtils.java
@@ -6,8 +6,8 @@ public final class DoubleUtils {
     private DoubleUtils() {
     }
 
-    public static double parseDoubleSecondDigit(double number) {
+    public static BigDecimal parseDoubleSecondDigit(double number) {
         BigDecimal bigDecimal = new BigDecimal(number);
-        return bigDecimal.setScale(2, BigDecimal.ROUND_HALF_UP).doubleValue();
+        return bigDecimal.setScale(2, BigDecimal.ROUND_HALF_UP);
     }
 }
